requirements.txt
appdirs
futures
jinja2
microdrop-plugin-manager>=0.5
path_helpers
pip_helpers>=0.6.post3
wheeler.pygtkhelpers>=0.14.post7
